backend/requirements.txt
# FastAPI y utilidades (las incluyo ya para no reescribir el fichero después)
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
pydantic==2.10.3
pydantic-settings==2.6.1

# Base de datos
sqlalchemy==2.0.36
psycopg[binary]==3.2.3
alembic==1.14.0

# Supabase (para Storage)
supabase==2.10.0

# Azure
openai==1.57.0
azure-search-documents==11.5.2
azure-core==1.32.0

# Procesamiento de documentos (los dejamos listos para después)
pypdf==5.1.0
python-docx==1.1.2
python-pptx==1.0.2
python-multipart==0.0.19